# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Waste Platform API",
    "description": "منصة إدارة النفايات الصلبة وإعادة التدوير — REST API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identifier",
                  "password"
                ],
                "properties": {
                  "identifier": {
                    "minLength": 3,
                    "maxLength": 120,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 1,
                    "maxLength": 128,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/auth/change-password": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "minLength": 1,
                    "maxLength": 128,
                    "type": "string"
                  },
                  "newPassword": {
                    "minLength": 8,
                    "maxLength": 128,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/auth/invitations/accept": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "password",
                  "displayName"
                ],
                "properties": {
                  "token": {
                    "minLength": 20,
                    "maxLength": 128,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 8,
                    "maxLength": 128,
                    "type": "string"
                  },
                  "displayName": {
                    "minLength": 2,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "phone": {
                    "minLength": 9,
                    "maxLength": 15,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/citizens/register": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "displayName",
                  "email",
                  "phone",
                  "password",
                  "idNumber",
                  "serviceAreaId"
                ],
                "properties": {
                  "displayName": {
                    "minLength": 2,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "email": {
                    "format": "email",
                    "maxLength": 120,
                    "type": "string"
                  },
                  "phone": {
                    "minLength": 9,
                    "maxLength": 15,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 8,
                    "maxLength": 128,
                    "type": "string"
                  },
                  "idNumber": {
                    "minLength": 9,
                    "maxLength": 9,
                    "type": "string"
                  },
                  "serviceAreaId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "addressHint": {
                    "maxLength": 160,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/admin/invitations": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "role"
                ],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 120,
                    "type": "string"
                  },
                  "role": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "collector"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "authority"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "sorter"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "finance"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "manager"
                        ]
                      }
                    ]
                  },
                  "serviceAreaId": {
                    "format": "uuid",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "type": "integer"
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "citizen"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "collector"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "authority"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "sorter"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "finance"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "manager"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "role",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/admin/invitations/{id}": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/admin/accounts": {
      "get": {
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "type": "integer"
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "citizen"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "collector"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "authority"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "sorter"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "finance"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "manager"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "role",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/admin/accounts/{id}/status": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "active"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "disabled"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/admin/audit": {
      "get": {
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "type": "integer"
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "citizen"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "collector"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "authority"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "sorter"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "finance"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "manager"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "role",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}